rec_gpu/requirements.txt
scikit-learn
dlib
face_recognition
opencv-contrib-python
setuptools
python-dotenv
gunicorn
eventlet
flask
flask-socketio
boto3
image